requirements.txt
flake8
pytest
bandit
python-dotenv
openai
pyttsx3